requirements.txt
httpx>=0.23.3
pillow>=9.2.0
pydantic>=1.10.0
sqlalchemy>=2.0.0
